requirements.txt
pywin32==300
python-dotenv>=0.19.2